+++ Sources/CodeEditTextView/TextView/TextView+Layout.swift
@@ -1,0 +1,95 @@
+//
+//  TextView+Layout.swift
+//  CodeEditTextView
+//
+//  Created by Khan Winter on 6/15/24.
+//
+
+import Foundation
+
+extension TextView {
+    open override class var isCompatibleWithResponsiveScrolling: Bool {
+        true
+    }
+
+    open override func prepareContent(in rect: NSRect) {
+        needsLayout = true
+        super.prepareContent(in: rect)
+    }
+
+    override public func draw(_ dirtyRect: NSRect) {
+        super.draw(dirtyRect)
+        if isSelectable {
+            selectionManager.drawSelections(in: dirtyRect)
+        }
+    }
+
+    override open var isFlipped: Bool {
+        true
+    }
+
+    override public var visibleRect: NSRect {
+        if let scrollView {
+            var rect = scrollView.documentVisibleRect
+            rect.origin.y += scrollView.contentInsets.top
+            return rect.pixelAligned
+        } else {
+            return super.visibleRect
+        }
+    }
+
+    public var visibleTextRange: NSRange? {
+        let minY = max(visibleRect.minY, 0)
+        let maxY = min(visibleRect.maxY, layoutManager.estimatedHeight())
+        guard let minYLine = layoutManager.textLineForPosition(minY),
+              let maxYLine = layoutManager.textLineForPosition(maxY) else {
+            return nil
+        }
+        return NSRange(
+            location: minYLine.range.location,
+            length: (maxYLine.range.location - minYLine.range.location) + maxYLine.range.length
+        )
+    }
+
+    public func updatedViewport(_ newRect: CGRect) {
+        if !updateFrameIfNeeded() {
+            layoutManager.layoutLines()
+        }
+        inputContext?.invalidateCharacterCoordinates()
+    }
+
+    @discardableResult
+    public func updateFrameIfNeeded() -> Bool {
+        var availableSize = scrollView?.contentSize ?? .zero
+        availableSize.height -= (scrollView?.contentInsets.top ?? 0) + (scrollView?.contentInsets.bottom ?? 0)
+        let newHeight = max(layoutManager.estimatedHeight(), availableSize.height)
+        let newWidth = layoutManager.estimatedWidth()
+
+        var didUpdate = false
+
+        if newHeight >= availableSize.height && frame.size.height != newHeight {
+            frame.size.height = newHeight
+            // No need to update layout after height adjustment
+        }
+
+        if wrapLines && frame.size.width != availableSize.width {
+            frame.size.width = availableSize.width
+            didUpdate = true
+        } else if !wrapLines && frame.size.width != max(newWidth, availableSize.width) {
+            frame.size.width = max(newWidth, availableSize.width)
+            didUpdate = true
+        }
+
+        if didUpdate {
+            needsLayout = true
+            needsDisplay = true
+            layoutManager.layoutLines()
+        }
+
+        if isSelectable {
+            selectionManager?.updateSelectionViews()
+        }
+
+        return didUpdate
+    }
+}
